Refactored out RoutineTasks into Task component via conditional rendering
