+++ client/src/components/Task.jsx
@@ -8,20 +8,45 @@ export const Task = ({ _id, name, type }) => {
     transform: CSS.Transform.toString(transform)
   }
 
+  const daysofWeek = ["M", "T", "W", "T", "F", "S", "S"];
+
   return (
     <>
-      {type === "wip" ? (
-          <div ref={setNodeRef} {...attributes} {...listeners} style={style} className="flex justify-around basis-2/3 h-5 w-125 rounded-sm border border-gray-400 shadow-md">
-            <div className="text-left min-w-30 max-w-30 place-self-center text-nowrap">
-              {name}
-            </div>
+      {type === "routine" ? (
+        <div className="flex justify-around rounded-sm border border-gray-400 shadow-md">
+          <div className="text-left min-w-30 max-w-30 place-self-center text-nowrap">
+            {name}
+          </div>
+          <div className="columns-1">
+            <p>Done?</p>
+            <input className="checked:accent-green-500/25" type="checkbox" />
+          </div>
+          <div className="columns-7">
+            {daysofWeek.map((day) => (
+              <div className="gap-2 justify-items-center">
+                <div>
+                  {day}
+                </div>
+                <input className="checked:accent-green-500/25" type="checkbox" />
+              </div>
+            ))}
+          </div>
+        </div>
+      ) : (
+        <div ref={setNodeRef} {...attributes} {...listeners} style={style} className="flex justify-around basis-2/3 h-5 w-125 rounded-sm border border-gray-400 shadow-md">
+
+          {type === "wip" ? (
+            <>
+              <div className="text-left min-w-30 max-w-30 place-self-center text-nowrap">
+                {name}
+              </div>
               <div className="columns-1 self-end">
                 <input className="checked:accent-green-500/25" type="checkbox" />
               </div>
-            </div>
-      ) : (
-        <div ref={setNodeRef} {...attributes} {...listeners} style={style} className="flex justify-around basis-2/3 h-5 w-125 rounded-sm border border-gray-400 shadow-md">
-            {name}
+            </>
+          ) : (
+            name
+          )}
         </div>
       )}
     </>
